connectComponent: byte to int16 (#150)
* Update contributing docs

* connectComponent: byte to int16

prep_aria.py and objects/stackDict.py:

+ change the data type of connectComponent dataset from BYTE to INT16 to support up to 32767 reliable regions/connected components for single interferograms.

+ turn ON compression of "lzf" while writing connectComponent to HDF5 file to save disk space

prep_aria.py:
+ add FILE_TYPE metadata
+ add MODIFICATION_TIME metadata to 3D datasets
+ turn ON chunks while writing HDF5 to speedup file IO
+ turn ON maxshape for 3D datasets to support updatable ifgramStack.h5: feature not used anyway yet; just to be consistent with mintpy.

update default plot settings for connectComponent dataset in view.py and  plot.py

+++ mintpy/objects/stack.py
@@ -555,7 +555,7 @@ Attributes         Dictionary for metadata
 /dropIfgram        1D array of bool    in size of (m,     ) with 0/False for drop and 1/True for keep
 /unwrapPhase       3D array of float32 in size of (m, l, w) in radian.
 /coherence         3D array of float32 in size of (m, l, w).
-/connectComponent  3D array of int8    in size of (m, l, w).           (optional)
+/connectComponent  3D array of int16   in size of (m, l, w).           (optional)
 /wrapPhase         3D array of float32 in size of (m, l, w) in radian. (optional)
 /rangeOffset       3D array of float32 in size of (m, l, w).           (optional)
 /azimuthOffset     3D array of float32 in size of (m, l, w).           (optional)
